Fix bug introduced by using `ref_install`
Changing the code to use `ref_install` to define where the installation
occurred caused a bug with installing custom reference genomes. This
turned out to be caused by the fact that some code was not used as a
module but as a stand-alone script, making the imports failing, but not
raising the correct error. Now fixed by calling the function within the
module.